+++ tests/unit/controls-filter.js
@@ -11,15 +11,15 @@ chai.use(require('chai-as-promised'));
 
 describe('Controls', () => {
     describe('Filter', () => {
-        let frag        = document.createDocumentFragment();
-        let container   = dom.getContainer();
-        let controls    = dom.getFilterControls();
+        const frag        = document.createDocumentFragment();
+        const container   = dom.getContainer();
+        const controls    = dom.getFilterControls();
 
         container.insertBefore(controls, container.children[0]);
 
         frag.appendChild(container);
 
-        let mixer = mixitup(container, {
+        const mixer = mixitup(container, {
             controls: {
                 scope: 'local'
             }
@@ -28,17 +28,17 @@ describe('Controls', () => {
         after(() => mixer.destroy());
 
         it('should detect nested filter controls and set active states upon instantiation', () => {
-            let filter = controls.querySelector('[data-filter="all"]');
+            const filter = controls.querySelector('[data-filter="all"]');
 
             chai.assert.isOk(filter.matches('.mixitup-control-active'));
         });
 
         it('should map filter controls with value "none" to the selector ""', () => {
-            let filter = controls.querySelector('[data-filter="none"]');
+            const filter = controls.querySelector('[data-filter="none"]');
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '');
             chai.assert.equal(state.totalShow, 0);
@@ -46,11 +46,11 @@ describe('Controls', () => {
         });
 
         it('should map filter controls with value "all" to the target selector', () => {
-            let filter = controls.querySelector('[data-filter="all"]');
+            const filter = controls.querySelector('[data-filter="all"]');
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '.mix');
             chai.assert.equal(state.totalHide, 0);
@@ -58,12 +58,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with a selector value', () => {
-            let filter = controls.querySelector('[data-filter=".category-a"]');
-            let totalMatching = container.querySelectorAll('.category-a').length;
+            const filter = controls.querySelector('[data-filter=".category-a"]');
+            const totalMatching = container.querySelectorAll('.category-a').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '.category-a');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -71,12 +71,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with a compound OR selector', () => {
-            let filter = controls.querySelector('[data-filter=".category-a, .category-b"]');
-            let totalMatching = container.querySelectorAll('.category-a, .category-b').length;
+            const filter = controls.querySelector('[data-filter=".category-a, .category-b"]');
+            const totalMatching = container.querySelectorAll('.category-a, .category-b').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '.category-a, .category-b');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -84,12 +84,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with a compound AND selector', () => {
-            let filter = controls.querySelector('[data-filter=".category-a.category-c"]');
-            let totalMatching = container.querySelectorAll('.category-a.category-c').length;
+            const filter = controls.querySelector('[data-filter=".category-a.category-c"]');
+            const totalMatching = container.querySelectorAll('.category-a.category-c').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '.category-a.category-c');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -97,12 +97,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with an attribute selector value', () => {
-            let filter = controls.querySelector('.mixitup_control__attr-a');
-            let totalMatching = container.querySelectorAll('[data-category="a"]').length;
+            const filter = controls.querySelector('.mixitup_control__attr-a');
+            const totalMatching = container.querySelectorAll('[data-category="a"]').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '[data-category="a"]');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -110,12 +110,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with an attribute selector value', () => {
-            let filter = controls.querySelector('.mixitup_control__attr-a');
-            let totalMatching = container.querySelectorAll('[data-category="a"]').length;
+            const filter = controls.querySelector('.mixitup_control__attr-a');
+            const totalMatching = container.querySelectorAll('[data-category="a"]').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '[data-category="a"]');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -123,12 +123,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with a compound OR attribute selector value', () => {
-            let filter = controls.querySelector('.mixitup_control__attr-a-or-b');
-            let totalMatching = container.querySelectorAll('[data-category="a"], [data-category="b"]').length;
+            const filter = controls.querySelector('.mixitup_control__attr-a-or-b');
+            const totalMatching = container.querySelectorAll('[data-category="a"], [data-category="b"]').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '[data-category="a"], [data-category="b"]');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -136,12 +136,12 @@ describe('Controls', () => {
         });
 
         it('should accept filter controls with a compound AND attribute selector value', () => {
-            let filter = controls.querySelector('.mixitup_control__attr-a-and-c');
-            let totalMatching = container.querySelectorAll('[data-category="a"][data-category="c"]').length;
+            const filter = controls.querySelector('.mixitup_control__attr-a-and-c');
+            const totalMatching = container.querySelectorAll('[data-category="a"][data-category="c"]').length;
 
             filter.click();
 
-            let state = mixer.getState();
+            const state = mixer.getState();
 
             chai.assert.equal(state.activeFilter.selector, '[data-category="a"][data-category="c"]');
             chai.assert.equal(state.totalShow, totalMatching);
@@ -149,31 +149,53 @@ describe('Controls', () => {
         });
 
         it ('should allow a single set of controls to control multiple mixer instance simultanously', () => {
-            let frag = document.createDocumentFragment();
+            const frag = document.createDocumentFragment();
 
-            let container1 = dom.getContainer();
-            let container2 = dom.getContainer();
-            let controls = dom.getFilterControls();
+            const container1 = dom.getContainer();
+            const container2 = dom.getContainer();
+            const controls = dom.getFilterControls();
 
             frag.appendChild(controls);
             frag.appendChild(container1);
             frag.appendChild(container2);
 
-            let mixer1 = mixitup(container1, {}, frag);
-            let mixer2 = mixitup(container2, {}, frag);
+            const mixer1 = mixitup(container1, {}, frag);
+            const mixer2 = mixitup(container2, {}, frag);
 
             after(() => {
                 mixer1.destroy();
                 mixer2.destroy();
             });
 
-            let filter = controls.querySelector('[data-filter=".category-a"]');
+            const filter = controls.querySelector('[data-filter=".category-a"]');
 
             filter.click();
 
             chai.assert.equal(mixer1.getState().activeFilter.selector, '.category-a');
             chai.assert.equal(mixer2.getState().activeFilter.selector, '.category-a');
             chai.assert.isOk(filter.matches('.mixitup-control-active'));
+        });
+
+        it ('should activate the appropriate controls on load for a single selector', () => {
+            const frag = document.createDocumentFragment();
+
+            const container = dom.getContainer();
+            const controls = dom.getFilterControls();
+
+            frag.appendChild(controls);
+            frag.appendChild(container);
+
+            const mixer = mixitup(container, {
+                load: {
+                    filter: '.category-a'
+                }
+            }, frag);
+
+            after(() => mixer.destroy());
+
+            const filter = controls.querySelector('[data-filter=".category-a"]');
+
+            chai.assert.isTrue(filter.classList.contains('mixitup-control-active'));
         });
     });
 });