+++ tests/unit/main.js
@@ -1,6 +1,7 @@
 require('./factory');
 require('./mixer-get-state');
 require('./mixer-filter');
+require('./mixer-toggle-on-off');
 require('./mixer-sort');
 require('./mixer-dataset');
 require('./mixer-insert');